Move iconer to common + casefix
