+++ src/common/components/ekspanderbart-innhold/ekspanderbartInnhold.less
