+++ src/common/components/ikoner/TerminIkon.tsx
